+++ NEWS.md
@@ -6,6 +6,8 @@ Version 1.1.1 - beta
 --------------------
 - Release notes: https://ezdxf.mozman.at/release-v1-1.html
 - NEW: Python 3.12 binary wheel deployment on PyPI
+- NEW: page alignment support in the `drawing` add-on for these backends: 
+  `SVGBackend`, `PyMuPdfBackend` and the `PlotterBackend`
 - INFO: `numpy` v1.25 has stopped providing Python 3.8 binary wheels on PyPI
 - BUGFIX: [#929](https://github.com/mozman/ezdxf/issues/929)
   handling of the minimum hatch line distance